Changed to make a simpler Hello World test
Mostly I've just removed all the peripheral setup, some of them
(like the Ethernet) wouldn't work properly without the chips fitted
and I wanted to just test the UART to make sure the MCU is working.
Also fixed a typo in the CMake that was using the wrong linker script
for the CM7 so its code was being placed in the wrong bank.